Use tuple instead of list for values of MappingListConnector and ArrayConnector
By allowing to define function input type as tuple instead of list, we allow users to decorate the relationship with `@functools.cache` decorator. 
Note that we also need to define immutable mapping as `dict` is not hashable.
My benchmark shows using `@functools.cache` decorator does not make execution faster, possibly because the computational cost of each relationship is relatively small.